+++ src/CommonDropdown.php
@@ -520,44 +520,67 @@ abstract class CommonDropdown extends CommonDBTM
     }
 
 
-    /** Check if the dropdown $ID is used into item tables
+    /**
+     * Check if the dropdown $ID is used into item tables
      *
      * @return boolean : is the value used ?
-     **/
+     */
     public function isUsed()
     {
         global $DB;
 
-        $ID = $this->fields['id'];
-
         $RELATION = getDbRelations();
-        if (isset($RELATION[$this->getTable()])) {
-            foreach ($RELATION[$this->getTable()] as $tablename => $field) {
-                if ($tablename[0] != '_') {
-                    if (!is_array($field)) {
-                        $row = $DB->request([
-                            'FROM'   => $tablename,
-                            'COUNT'  => 'cpt',
-                            'WHERE'  => [$field => $ID]
-                        ])->current();
-                        if ($row['cpt'] > 0) {
-                             return true;
-                        }
+
+        if (!array_key_exists($this->getTable(), $RELATION)) {
+            return false;
+        }
+
+        foreach ($RELATION[$this->getTable()] as $tablename => $fields) {
+            if ($tablename[0] == '_') {
+                continue; // Ignore relations prefxed by `_`
+            }
+
+            $or_criteria = [];
+
+            foreach ($fields as $field) {
+                if (is_array($field)) {
+                    // Relation based on 'itemtype'/'items_id' (polymorphic relationship)
+                    if ($this instanceof IPAddress && in_array('mainitemtype', $field) && in_array('mainitems_id', $field)) {
+                        // glpi_ipaddresses relationship that does not respect naming conventions
+                        $itemtype_field = 'mainitemtype';
+                        $items_id_field = 'mainitems_id';
                     } else {
-                        foreach ($field as $f) {
-                             $row = $DB->request([
-                                 'FROM'   => $tablename,
-                                 'COUNT'  => 'cpt',
-                                 'WHERE'  => [$f => $ID]
-                             ])->current();
-                            if ($row['cpt'] > 0) {
-                                return true;
-                            }
-                        }
+                        $itemtype_matches = preg_grep('/^itemtype/', $field);
+                        $items_id_matches = preg_grep('/^items_id/', $field);
+                        $itemtype_field = reset($itemtype_matches);
+                        $items_id_field = reset($items_id_matches);
                     }
+                    $or_criteria[] = [
+                        $itemtype_field => $this->getType(),
+                        $items_id_field => $this->getID(),
+                    ];
+                } else {
+                    // Relation based on single foreign key
+                    $or_criteria[] = [
+                        $field => $this->getID(),
+                    ];
                 }
             }
+
+            if (count($or_criteria) === 0) {
+                return false;
+            }
+
+            $row = $DB->request([
+                'FROM'   => $tablename,
+                'COUNT'  => 'cpt',
+                'WHERE'  => ['OR' => $or_criteria]
+            ])->current();
+            if ($row['cpt'] > 0) {
+                 return true;
+            }
         }
+
         return false;
     }