+++ src/Computer.php
@@ -310,26 +310,14 @@ class Computer extends CommonDBTM
 
     public function cleanDBonPurge()
     {
-
         $this->deleteChildrenAndRelationsFromDb(
             [
-                Certificate_Item::class,
                 Computer_Item::class,
-                Item_SoftwareLicense::class,
-                Item_SoftwareVersion::class,
-                Item_Process::class,
-                Item_Environment::class,
                 ComputerAntivirus::class,
                 ComputerVirtualMachine::class,
-                Item_Disk::class,
-                Item_Project::class,
+                Item_Environment::class,
+                Item_Process::class,
             ]
-        );
-
-        Item_Devices::cleanItemDeviceDBOnItemDelete(
-            $this->getType(),
-            $this->fields['id'],
-            (!empty($this->input['keep_devices']))
         );
     }