+++ src/Dashboard/Right.php
@@ -41,6 +41,7 @@ class Right extends \CommonDBChild
     public static $items_id = 'dashboards_dashboards_id';
 
    // prevent bad getFromDB when bootstraping tests suite
+   // FIXME Should be true
     public static $mustBeAttached = false;
 
     /**